app/requirements.txt
Django==3.1.1
psycopg2-binary==2.8.5
django-bootstrap-modal-forms==2.0.0
django-widget-tweaks=1.4.8